next

+++ src/main/java/com/isc/astd/service/UserService.java
@@ -2,11 +2,11 @@ package com.isc.astd.service;
 
 import com.isc.astd.domain.Audit;
 import com.isc.astd.domain.User;
+import com.isc.astd.repository.DocRepository;
+import com.isc.astd.repository.FileRepository;
 import com.isc.astd.repository.UserRepository;
-import com.isc.astd.service.dto.PageRequestDTO;
-import com.isc.astd.service.dto.PageableDTO;
-import com.isc.astd.service.dto.UserBaseDTO;
-import com.isc.astd.service.dto.UserDTO;
+import com.isc.astd.repository.specification.UserSpecification;
+import com.isc.astd.service.dto.*;
 import com.isc.astd.service.mapper.Mapper;
 import org.springframework.context.annotation.Lazy;
 import org.springframework.data.domain.Page;
@@ -14,6 +14,8 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.time.LocalDate;
 
 /**
  * @author p.dzeviarylin
@@ -25,13 +27,17 @@ public class UserService {
     private final UserRepository userRepository;
     private final PositionService positionService;
     private final CatalogService catalogService;
+    private final FileRepository fileRepository;
+    private final DocRepository docRepository;
 
     private final Mapper mapper;
 
-    public UserService(UserRepository userRepository, @Lazy PositionService positionService, @Lazy CatalogService catalogService, Mapper mapper) {
+    public UserService(UserRepository userRepository, @Lazy PositionService positionService, @Lazy CatalogService catalogService, FileRepository fileRepository, DocRepository docRepository, Mapper mapper) {
         this.userRepository = userRepository;
 		this.positionService = positionService;
 		this.catalogService = catalogService;
+		this.fileRepository = fileRepository;
+		this.docRepository = docRepository;
 		this.mapper = mapper;
     }
 
@@ -45,9 +51,10 @@ public class UserService {
         return mapper.map(user, UserDTO.class);
     }
 
-	public PageRequestDTO<UserDTO> getAllUsers(org.springframework.security.core.userdetails.User principal, PageableDTO pageableDTO) {
+	public PageRequestDTO<UserDTO> getAllUsers(org.springframework.security.core.userdetails.User principal, PageableDTO pageableDTO, UserFilterDTO filterDTO) {
 		Page<User> users = userRepository.findAll(
-				PageRequest.of(pageableDTO.getPage() - 1, pageableDTO.getLimit(), new Sort(Sort.Direction.DESC, "id"))
+				UserSpecification.filter(filterDTO),
+				PageRequest.of(pageableDTO.getPage() - 1, pageableDTO.getLimit(), Sort.by(Sort.Direction.DESC, "id"))
 		);
 		return new PageRequestDTO<>(users.getTotalPages(), users.getTotalElements(), mapper.mapAsList(users, UserDTO.class));
 	}
@@ -66,7 +73,11 @@ public class UserService {
 
 	public UserBaseDTO updateUser(String id, UserBaseDTO dto) {
 		User user;
-    	if(!id.equals(dto.getPrevId())){
+    	if(!id.equals(dto.getPrevId())) {
+    		if(fileRepository.countAllByCreatedByEqualsOrLastModifiedByEquals(dto.getPrevId(), dto.getPrevId()) > 0 ||
+					docRepository.countAllByCreatedByEqualsOrLastModifiedByEquals(dto.getPrevId(), dto.getPrevId()) > 0) {
+    			throw new RuntimeException("Нельзя изменить ID пользователя");
+			}
 			deleteUser(dto.getPrevId()); // it's not possible to change PK id
 			user = mapper.map(dto, User.class);
 		} else {
@@ -88,8 +99,10 @@ public class UserService {
 
 	public void deleteUser(String id) {
 		User user = getUser(id);
+		user.setExpiredDate(LocalDate.now());
 		user.setAuditAction(Audit.Action.USER_DELETE);
-		userRepository.delete(user);
+//		userRepository.delete(user);
+		userRepository.save(user);
 	}
 
 	public UserBaseDTO getUserById(String id) {
